Added: Interessen Administration: List, Edit, Add, Delete Interessen & Fix in User Administration
+ renamed backend UserController to A_UserController
+ EventsController + basic view
+ InteressenController + view that lists interessen
+ add, edit, delete interessen 
+ edit interessen via modal
+ searchable list

+ fixed in user administration: error when no interest selected

+++ app/backend/models/A_HasInterestModel.php
@@ -53,7 +53,8 @@ class A_HasInterestModel extends Model {
      * @return void
      * set the given interests for the given user
      */
-    public function setInterestsForUserID(int $UserID, array $interests) {
+    public function setInterestsForUserID(int $UserID, $interests) {
+        if(is_null($interests)) return;
         $sql = "DELETE FROM hasInterest WHERE id_user = :user_id";
         $statement = $this->db->prepare($sql);
         $statement->bindValue(':user_id', $UserID);
